add QM feature flag checks on ClaimSettledDefendantDashboardNotificationHandler

+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/dashboardnotifications/defendant/ClaimSettledDefendantDashboardNotificationHandler.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.callback.DashboardCallbackHandler;
+import uk.gov.hmcts.reform.civil.enums.YesOrNo;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.service.DashboardNotificationsParamsMapper;
 import uk.gov.hmcts.reform.civil.service.FeatureToggleService;
@@ -53,15 +54,18 @@ public class ClaimSettledDefendantDashboardNotificationHandler extends Dashboard
     @Override
     protected void beforeRecordScenario(CaseData caseData, String authToken) {
         final String caseId = String.valueOf(caseData.getCcdCaseReference());
-        dashboardNotificationService.deleteByReferenceAndCitizenRole(
-            caseId,
-            "DEFENDANT"
-        );
+        if(featureToggleService.isQueryManagementLRsEnabled() && ( caseData.getEaCourtLocation().equals(YesOrNo.YES))){
+            dashboardNotificationService.deleteByReferenceAndCitizenRole(
+                caseId,
+                "DEFENDANT"
+            );
 
-        taskListService.makeProgressAbleTasksInactiveForCaseIdentifierAndRole(
-            caseId,
-            "DEFENDANT",
-            null
-        );
+            taskListService.makeProgressAbleTasksInactiveForCaseIdentifierAndRole(
+                caseId,
+                "DEFENDANT",
+                null
+            );
+        }
+
     }
 }
